unify data layout
