+++ paddle/fluid/framework/data_layout.h
@@ -18,58 +18,4 @@ limitations under the License. */
 #include <ostream>
 #include <string>
 
-#include "paddle/fluid/platform/enforce.h"
-
-namespace paddle {
-namespace framework {
-
-enum class DataLayout {
-  kNHWC = 0,
-  kNCHW = 1,
-  kAnyLayout = 2,
-  kMKLDNN = 3,  // all layouts supported by MKLDNN internally
-};
-
-inline DataLayout StringToDataLayout(const std::string& str) {
-  std::string s(str);
-  for (size_t i = 0; i < s.size(); ++i) {
-    s[i] = toupper(s[i]);
-  }
-
-  if (s == "NHWC") {
-    return DataLayout::kNHWC;
-  } else if (s == "NCHW") {
-    return DataLayout::kNCHW;
-  } else if (s == "ANYLAYOUT") {
-    return DataLayout::kAnyLayout;
-  } else if (s == "MKLDNNLAYOUT") {
-    return DataLayout::kMKLDNN;
-  } else {
-    PADDLE_THROW(platform::errors::InvalidArgument(
-        "Unknown data layout type string: %s.", s));
-  }
-}
-
-inline std::string DataLayoutToString(const DataLayout& data_layout) {
-  switch (data_layout) {
-    case DataLayout::kNHWC:
-      return "NHWC";
-    case DataLayout::kNCHW:
-      return "NCHW";
-    case DataLayout::kAnyLayout:
-      return "ANY_LAYOUT";
-    case DataLayout::kMKLDNN:
-      return "MKLDNNLAYOUT";
-    default:
-      PADDLE_THROW(platform::errors::InvalidArgument(
-          "Unknown Data Layout type %d.", data_layout));
-  }
-}
-
-inline std::ostream& operator<<(std::ostream& out, const DataLayout& l) {
-  out << DataLayoutToString(l);
-  return out;
-}
-
-}  // namespace framework
-}  // namespace paddle
+#include "paddle/pten/common/layout.h"
